ai_task_planner/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-task-planner"
version = "0.1.0"
description = "Desktop AI Task Planner & Google Calendar sync application"
readme = "README.md"
authors = [{ name = "AI Assistant" }]
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6",
    "SQLAlchemy>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.25",
    "google-api-python-client>=2.100",
    "google-auth>=2.22",
    "google-auth-oauthlib>=1.0",
    "google-auth-httplib2>=0.1",
    "apscheduler>=3.10",
    "plyer>=2.1",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.11",
    "mypy>=1.8",
    "ruff>=0.1.6",
]

[project.scripts]
ai-task-planner = "src.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B", "Q", "S", "BLE", "C4", "SIM", "PL", "RUF"]
ignore = ["S101", "BLE001"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true
mypy_path = "src"
explicit_package_bases = true
plugins = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["src/tests"]
